Convert black to currentColor in icons using SVGO
We were previously styling the symbols to set fill='black' to fill='currentColor' via CSS, but this change was not inherited by instances of the icons in Firefox, as it was in Safari and Chrome.

This technique of rewriting the color values in the SVGs in the build will be much more reliable.