Hide SVG spritesheet from accessibility tools
VoiceOver for iOS was reading out the titles of all SVG sprites.

The API for svg-sprite-loader to customize the attributes that are applied to
the SVG spritesheet's svg element is not great; requiring us to provide a custom
sprite module that reimplements the logic for injecting the spritesheet into the
page. Hopefully svg-sprite-loader will improve this API in a future release.