Refactor Kebab component to use an independent MenuList
Reason we didn't like the old design: it encouraged a developer who was using the component to pass in the name of a CSS class for the header letting them mess with the internal styling, and the whole design had a lot of assumptions (eg header colours) that were quite specific to the survey browser.

The new design allows you to pass in any component to the Kebab, and all the Kebab does is handle the opening, closing and positioning.

The MenuList class and it's children (MenuItem, MenuHeader and MenuSeparator) can then be mixed and matched as desired. You can also create your own custom MenuHeader with custom styling, which is what we would do for the survey-browser use case.