Use CSS module composition for kebab hover styling
Previously I was reaching in to style an `svg` element but that's bad form as we might change the internal representation in future.

Now I have a separate `kebabHoverArea` class that will turn on the hover functionality for the kebab, and the container you want to be the hover area can compose this class.

Note: I ran into a glitch with using compose here. The workaround was to upgrade style-loader. See https://github.com/gatsbyjs/gatsby/pull/3283/