Update active/passive voice section to be accessible
The confluence article relied on colour to differentiate the actor, action and thing being acted upon. Using <em> <b> <u> has the semantic meaning of "display differently" without giving any hint as to meaning of the different display.  It also happens to be much easier for a colour-blind person like me to make sense of.