Workaround for webpack1 SVGO handling white colours
The SVGO config had a regular expression in it, and because Gatsby uses Webpack 1, and because Webpack 1 config is sent as JSON strings, regular expressions don't work.

I've worked around this by only replacing `#000000` with `currentColor`, which is good enough for now, and is a hack we can undo if we ever get Gatsby running with Webpack 2.