Fix for NavigationBar correctly picking Link components
For some reason I don't understand, testing against the actual type is working on murmur but not on Gatsby. I tried switching to test the type name, which worked on both, but I didn't realise that when the code is minified, the name changes, and so they're no longer correctly identified.

This fixes that problem.

Interestingly, this wouldn't be caught by visual regression tests, because they run with unminified code.