Use our own loader for adding postcss support.
This is required because gatsby-plugin-postcss-sass will use css-loader with the setting `importLoaders: 1` (this is non-configurable), and as a result when you use `compose: x from 'y.module.scss'` it will not use both postcss loader and sass loader, and then will start having issues. Painful!

One option that might be fun to explore is using postcss-loader modules rather than css-loader modules, which would give us the option to run NodeJS and output some files when we process. Notably, we could output Elm/FlowType files so we start strictly typing our CSS class names.