Various fixes to get the site mostly working again with correct fonts.
There are still issues with some spacing and sizing, we're not yet properly using the baseline grid, and I'm willing to bet a bunch of stuff (spacing, padding, sizing) that was pixel perfect before is now different. But at least it's consistently applied now and the worst problems are dealt with.