Get text content sitting nicely on the baseline grid
Now, other than images, everything on the product language page
sits nicely on the grid (or if not the main grid lines, then the
12px half grid lines)

Note: I've been testing in FF and it's almost perfect there, but
things are still drifting up from the baseline in Safari/Chrome.
Will need to investigate further.