Correct the positioning of the dropdown menus.
The dropdown menus were previously positioned absolutely relative to the page. Unfortunately the new type styles apply a `transform: translateY()` to the color block, which seems to count as giving an element "position", so now absolute positioning would be relative to the block instead, meaning the JS values will be completely wrong.

The workaround I've done is to absolutely position it relative to the kebab icon, and this works quite well.

There is a remaining issue with z-indexes though: because we have `transform: translateY()` used on many elements, we have a tonne of different stacking contexts, so merely setting a z-index does not seem to be helping. Will investigate this in the next commit.