Use `position: relative; top: $offset` for baseline transforms
Previously it was `transform: translateY($offset)`. The problem with this approach is that the transform creates a stacking context and breaks our popup menus and other z-indexed items.

I've fixed a few places where this changed layout (mostly it was cancelling out the baseline shift to prevent nested elements from being shifted twice).

The icon and colours pages are looking good, but I think the language page is demonstrating a slight drift that will need to be corrected.